requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==41.0.7
Django==4.2
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-resized==1.0.2
gunicorn==21.2.0
idna==3.6
packaging==23.2
Pillow==10.1.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.27.1
shippo==2.1.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.4
stripe==7.12.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
django-cleanup==8.1.0
gunicorn